[test] Factor WidthHelpers out of ChiselSpec
Move the WidthHelpers trait out of ChiselSpec and into the one file that
uses it.  This makes it easier to keep that test self-contained.
