+++ src/test/scala-2/chiselTests/WidthSpec.scala
@@ -3,6 +3,10 @@
 package chiselTests
 
 import chisel3._
+import chisel3.reflect.DataMirror
+import circt.stage.ChiselStage
+import org.scalactic.source.Position
+import org.scalatest.Assertions
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 
@@ -44,6 +48,58 @@ class WidthSpec extends AnyFlatSpec with Matchers {
     0.U.getWidth should be(1)
     0.S.getWidth should be(1)
   }
+}
+
+trait WidthHelpers extends Assertions {
+
+  def assertKnownWidth(expected: Int, args: Iterable[String] = Nil)(gen: => Data)(implicit pos: Position): Unit = {
+    class TestModule extends Module {
+      val testPoint = gen
+      assert(testPoint.getWidth === expected)
+      val out = IO(chiselTypeOf(testPoint))
+      // Sanity check that firrtl doesn't change the width
+      val zero = 0.U(0.W).asTypeOf(chiselTypeOf(testPoint))
+      if (DataMirror.isWire(testPoint)) {
+        testPoint := zero
+      }
+      out := zero
+      out := testPoint
+    }
+    val verilog = ChiselStage.emitSystemVerilog(new TestModule, args.toArray, Array("-disable-all-randomization"))
+    expected match {
+      case 0 => assert(!verilog.contains("out"))
+      case 1 =>
+        assert(verilog.contains(s"out"))
+        assert(!verilog.contains(s"0] out"))
+      case _ => assert(verilog.contains(s"[${expected - 1}:0] out"))
+    }
+  }
+
+  def assertInferredWidth(expected: Int, args: Iterable[String] = Nil)(gen: => Data)(implicit pos: Position): Unit = {
+    class TestModule extends Module {
+      val testPoint = gen
+      assert(!testPoint.isWidthKnown, s"Asserting that width should be inferred yet width is known to Chisel!")
+      // Sanity check that firrtl doesn't change the width
+      val widthcheck = Wire(chiselTypeOf(testPoint))
+      dontTouch(widthcheck)
+      val zero = 0.U(0.W).asTypeOf(chiselTypeOf(testPoint))
+      if (DataMirror.isWire(testPoint)) {
+        testPoint := zero
+      }
+      widthcheck := zero
+      widthcheck := testPoint
+    }
+    val verilog =
+      ChiselStage.emitSystemVerilog(new TestModule, args.toArray :+ "--dump-fir", Array("-disable-all-randomization"))
+    expected match {
+      case 0 => assert(!verilog.contains("widthcheck"))
+      case 1 =>
+        assert(verilog.contains(s"widthcheck"))
+        assert(!verilog.contains(s"0] widthcheck"))
+      case _ => assert(verilog.contains(s"[${expected - 1}:0] widthcheck"))
+    }
+  }
+
 }
 
 abstract class WireRegWidthSpecImpl extends AnyFlatSpec with Matchers with WidthHelpers {
